requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
pydantic
PyMuPDF
